chore(app): tidy up FAQ
